finnaly video streming is implement in backend

+++ server/src/utils/cloudinay.js
@@ -1,38 +1,41 @@
-import {v2 as cloudinary} from "cloudinary"
-import fs from "fs"
-import  dotenv from "dotenv";
+import { v2 as cloudinary } from "cloudinary";
+import fs from "fs";
+import dotenv from "dotenv";
 
-dotenv.config()
+dotenv.config();
 
-
-cloudinary.config({ 
-  cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
-  api_key: process.env.CLOUDINARY_API_KEY, 
-  api_secret: process.env.CLOUDINARY_API_SECRAT 
+cloudinary.config({
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRAT,
 });
 
-
-
 const uploadOnCloudinary = async (localFilePath) => {
-    try {
-        if (!localFilePath) return null
-        // console.log(localFilePath);
-        
-        const response = await cloudinary.uploader.upload(localFilePath, {
-            resource_type: "auto"
-        })
-        
-        fs.unlinkSync(localFilePath)
-        return response;
+  try {
+    if (!localFilePath) return null;
+    // console.log(localFilePath);
 
-    } catch (error) {
-      console.log(error);
-      
-        fs.unlinkSync(localFilePath)
-        return null;
+    const response = await cloudinary.uploader.upload(localFilePath, {
+      resource_type: "auto",
+    });
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
     }
-}
 
+    return response;
+    
+  } catch (error) {
 
+    if (fs.existsSync(localFilePath)) {
+      try {
+        fs.unlinkSync(localFilePath);
+      } catch (e) {
+        console.warn("Failed to delete file in catch:", e.message);
+      }
+    }
 
-export {uploadOnCloudinary}
+    return null;
+  }
+};
+
+export { uploadOnCloudinary };
